requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
numpy==1.19.4
paho-mqtt==1.5.1
pandas==1.1.4
pkg-resources==0.0.0
protobuf==3.14.0
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
Werkzeug==1.0.1
